rotate-password/requirements.txt
mysql-connector-python
bcrypt
qrcode
Pillow
